What game should we create?
In college I made a dice game, I have to open my zip file to confirm. But I am aware that people also make card games or fill-in-the-blank games. All are fine with me! Should we add players? Bobby and Aurelie follows me on Github but neither have reached out to me yet.